[CSL-778] Move part of Balances to Pos.DB.GState.Balances, move maybeThrow to core
