+++ src/Pos/Context/Functions.hs
@@ -34,7 +34,8 @@ import           Pos.Context.Context     (NodeContext (..), ncGenesisLeaders,
 import           Pos.Lrc.Error           (LrcError (..))
 import           Pos.Txp.Toil.Types      (Utxo)
 import           Pos.Types               (EpochIndex, HeaderHash, SlotLeaders)
-import           Pos.Util                (maybeThrow, readTVarConditional)
+import           Pos.Util                (readTVarConditional)
+import           Pos.Util           (maybeThrow)
 
 ----------------------------------------------------------------------------
 -- Genesis
