+++ src/Pos/Ssc/GodTossing/GState.hs
@@ -1,5 +1,5 @@
-{-# LANGUAGE Rank2Types          #-}
-{-# LANGUAGE TypeFamilies        #-}
+{-# LANGUAGE Rank2Types   #-}
+{-# LANGUAGE TypeFamilies #-}
 
 -- | Instance of SscGStateClass.
 
@@ -46,9 +46,8 @@ import           Pos.Types                      (Block, EpochIndex (..), SlotId 
                                                  blockMpc, epochIndexL, epochOrSlotG,
                                                  gbHeader)
 import           Pos.Util                       (NE, NewestFirst (..), OldestFirst (..),
-                                                 maybeThrow, toOldestFirst, _neHead,
-                                                 _neLast)
-
+                                                 toOldestFirst, _neHead, _neLast)
+import           Pos.Util                  (maybeThrow)
 ----------------------------------------------------------------------------
 -- Utilities
 ----------------------------------------------------------------------------
