+++ src/Pos/Txp/Toil/DBTxp.hs
@@ -26,6 +26,7 @@ import           Universum
 import           Pos.Context                 (WithNodeContext)
 import           Pos.DB.Class                (MonadDB)
 import qualified Pos.DB.GState               as GS
+import qualified Pos.DB.GState.Balances      as GS
 import           Pos.Delegation.Class        (MonadDelegation)
 import           Pos.Ssc.Extra               (MonadSscMem)
 import           Pos.Update.MemState.Class   (MonadUSMem (..))
